Decompose scrolling strips
