+++ src/common/mod.rs
@@ -12,6 +12,7 @@ mod resize_mode;
 mod resize_state;
 mod sizing;
 mod transient_workspace_id;
+mod width_preset;
 mod window;
 mod window_handle;
 mod window_placement;
@@ -32,6 +33,7 @@ pub use crate::common::resize_mode::ResizeMode;
 pub use crate::common::resize_state::ResizeState;
 pub use crate::common::sizing::Sizing;
 pub use crate::common::transient_workspace_id::TransientWorkspaceId;
+pub(crate) use crate::common::width_preset::WidthPreset;
 pub use crate::common::window::Window;
 pub use crate::common::window_handle::WindowHandle;
 pub use crate::common::window_placement::WindowPlacement;
